requirement.txt
gradio==5.24.0
numpy==1.26.4
torch>=2.0.0
pillow>=10.0.0
transformers>=4.34.0
scikit-learn>=1.2.0
pandas>=2.0.0
matplotlib>=3.0.0
seaborn>=0.12.0